Calculator to double

+++ src/main/java/org/example/Calculator.java
@@ -2,7 +2,7 @@ package org.example;
 
 public class Calculator {
 
-    public static int add(int number1, int number2)
+    public static double add(double number1, double number2)
     {
         return number2+number1;
     }
